fixed nits and added folder `ReusableStyles`
